keep index.js and add another dynamic page  at root
